refactor(domain): move INVALID_PURCHASE_UNIT from presentation to domain
Update PurchaseAmount and its tests to import the constant from lotto.domain.ErrorMessages.
