+++ src/main/kotlin/lotto/domain/ErrorMessages.kt
@@ -1,0 +1,5 @@
+package lotto.domain
+
+object ErrorMessages {
+    const val INVALID_PURCHASE_UNIT = "[ERROR] Purchase amount must be in multiples of 1,000 KRW."
+}
