+++ src/main/kotlin/lotto/presentation/ErrorMessages.kt
@@ -2,5 +2,4 @@ package lotto.presentation
 
 object ErrorMessages {
     const val INVALID_NUMBER = "[ERROR] Input is not a number."
-    const val INVALID_PURCHASE_UNIT = "[ERROR] Purchase amount must be in multiples of 1,000 KRW."
 }
